lnet: Discard pt_zombies_list
When an lnet_peer_ni is removed from the hash table, it is added to a
pt_zombies_list.  The only time the content of this list is examined
is to assert that it is empty.  As we keep a counter of the number
of zombies (pt_zombies) we can equally well assert that the count is
zero.

So do that, and stop maintaining the list.

Signed-off-by: NeilBrown <neilb@suse.com>