lustre: lnet: convert ni_refs to percpu_refcount.
ni_refs is a per-cpt refcount.
Linux already has a per-cpu refcount implementation
which doesn't require any locking.

So convert ni_refs to percpu_refcount.
As a bonus, we can get a wake-up when the refcount
reaches zero, rather than having to wait a full second.
The waiting in lnet_clear_zombies_nis_locked() is
modified so that instead of waiting one second each
time, and printing a warning on power-of-two seconds,
we wait an increasing power-of-two seconds and print
a warning if the wait ever timed out.

Signed-off-by: NeilBrown <neilb@suse.com>