lnet: prepare for RCU lookup of struct lnet_ni
struct lnet_ni is initially on a ni_netlist list.  When it is removed,
it is moved to a zombie list.  This pattern doesn't work very well for
RCU lookup as the list .next pointer mustn't change.
So change the zombie list to be singly linked, using the
ni_netlist.prev link.

Also use kfree_rcu() to perform the final free.

With this, it is safe to walk the net_ni_list of lnet_ni with only RCU
protection.

Signed-off-by: NeilBrown <neilb@suse.com>