lnet: treat ln_nets as empty when LNET_STATE_STOPPING
We currently move all lnet_nets to a zombie list
when stopping the lnet.
This ensures no new references are taken.

Doing this makes it harder to use RCU to protect
the walking of the list as the move might happen
during a walk, and then the end will never be found.

So change all the walking code (that is not already protected by
ln_api_mutex) to check ln_state *after* finding something on
the list.  If ln_state is not LNET_STATE_RUNNING, then
act as though the list were empty.

Use smp_store_release() to clear the RUNNING state, and
smp_load_acquire() to check it, to ensure proper barriers.

Now we can remove the zombie list, and just deleted everything on
ln_nets one by one.

Signed-off-by: NeilBrown <neilb@suse.com>