lnet: change lnet_get_next_ni_lock to use rcu protection.
Now that the lnet_net and lnet_ni lists can be walked
with only RCU protection, change lnet_get_next_ni_locked() to
use that protection and rename it to lnet_get_next_ni_rcu().

This involves using list_first_or_null_rcu() to both
check if there is a first entry on the list, and return it.

Also change all callers of this function to provide rcu protection
with rcu_read_lock().

In general, a reference must be taken to the lnet_ni before dropping
the rcu lock.  In one place (lnet_get_best_ni()) we cannot do that
yet as the callers don't expect it to return a counted reference.
As long as this is resolved before we remove the spinlocking, this is
safe.

Also change lnet_is_ni_healthy_locked() to lnet_is_ni_healthy()
as it doesn't need any locking, and change
lnet_update_ni_status_locked() to lnet_update_ni_status() as the
addition of RCU protection means no other lock is needed.

Signed-off-by: NeilBrown <neilb@suse.com>