lnet: convert lnet_get_net_locked() to lnet_get_net_rcu()
We no longer need a spinlock to safely walk the lists to find an
lnet_net, we can use RCU.
There is no refcount we can increment to get a stable pointer to
a lnet_net, so require that if the pointer needs to be dereferenced,
either the rcu read lock or the api_mutex must be held.
No called that depends on RCU needs the pointer for very long,
so this is safe.

Signed-off-by: NeilBrown <neilb@suse.com>