lnet: discard lnet_{nid,net}2ni_locked()
Now that we can walk the relevant lists using RCU,
lnet_nid2ni_addref() and lnet_net2ni_addref() can be implemented
without spinlocking and lnet_nid2ni_locked() and
lnet_net2ni_locked() can be discarded.

We need the addref version as we need a counted reference to keep the
ni stable after dropping the rcu read lock.  This reference is taken
with a "tryget" as it could race with the ni being freed.

This change particularly affects lnet_select_pathway() as 'best_ni'
is now always a counted reference.  So we need to decref when we
return or "goto again" without using the reference.

This change will eventually allow some locking to be removed.

Signed-off-by: NeilBrown <neilb@suse.com>