lnet: use RCU for remaining walks of net and ni lists.
When we walk the net and ni lists with the api_mutex held, we
don't need RCU protection, so those walks aren't changed.
Also when walking a list that isn't shared, such as a temporary list
that was built while configuring a network and is now being destroyed
because the config failed - in that case no further protection is
needed.

But in all cases where api_mutex isn't held and the list might be
shared, RCU is now used to protect the walking of lists of nets and
nis.

This mean that we no longer need the percpt lock to change these
lists, the api_mutex is enough.

Note there is a small optimization added in LNetGetId() - a 'break'
appear to be intended to abort two nested loops, but actually only
aborts one.  This is not a correctness problem, but it looks odd.
Add a test to break the outer loop too.

Signed-off-by: NeilBrown <neilb@suse.com>