lnet: discard lnd_refcount
The lnd_refcount in 'struct lnet_lnd' is never tested (except in an
ASSERT()), so it cannot be needed.  Let's remove it.

Each individual lnd keeps track of how many lnet_ni are registered for
that lnd e.g.  ksocklnd has a counter in ksnd_nnets and o2iblnd has a
linked list in kib_devs.  They hold a reference on the module while
there are registered devices, and the lnd is only freed (and the
lnd_refcount checked) when the module is unloaded.  This confirms that
lnd_refcount adds no value.

Signed-off-by: NeilBrown <neilb@suse.de>