lnet: improve locking on lnd registration/deregistration.
It is important to hold the lnd_mutex on lnd startup and shutdown.
The lnd module manages the refcount of the module and without
holding the mutex at these times, the module could get unloaded
while in lnd code.

Also remove the taking of lnd_mutex from lnet_find_lnd_by_type().
This was helpful when there was a matching lnd_put(), but now
that it gone, there is little to be gained.

Signed-off-by: NeilBrown <neilb@suse.de>