lustre: llite: Don't access lov_md fields before size check
When 'struct lov_user_md' is passed in via setxattr, it comes with
a size.  If that size is too small, some functions that check exactly
what version is present might access beyond the end of allocated
memory, which can have undesirable effects, such as triggering
a KASAN warning (and possibly worse).

So check that the size is sane before looking inside the structure
at all.

Signed-off-by: NeilBrown <neilb@suse.de>