staging: mt7621-dts: Fix LAN LEDs on gb-pc2
The LEDs on the LAN ports on the Gnubee-PC2 are wired
differently than on the PC1.
On the PC2, the I2C pins are used to drive these
two LEDs.

So update the dts file accordingly.

This *doesn't* fix the LEDs on the third (yellow) port.
It is quite separate.

Signed-off-by: NeilBrown <neilb@suse.de>