staging: mt7621-dts: possible fix for LAN3 LEDs
The third lan on the gnubee-pc2 has, according to the schmatic, both leds
connected to GPIOs - 13 and 15.
I cannot make them work on my board, but I have a pre-release board, so
maybe something is wrong.

These two pins are part of the JTAG group, so we need to set that
to 'gpio' mode, and declare the two gpios as driving LEDs.
Maybe it will work on other boards.

Signed-off-by: NeilBrown <neil@brown.name>