lustre: obdecho: use vmalloc for lnb
When allocating the niobuf_local, if there are a large number of
(potential) fragments this allocation can be quite large. Use
kvmalloc_array() and kvfree() to avoid allocation errors and
console noise. This was causing sanity test_180c to fail in a
VM on occasion, and could also be problem in real use.

WC-bug-id: https://jira.whamcloud.com/browse/LU-10903
Lustre-commit: 8878bab7ae5f ("LU-10903 obdecho: use OBD_ALLOC_LARGE for lnb")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/31964
Reviewed-by: Emoly Liu <emoly@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>