lnet: libcfs: poll fail_loc in cfs_fail_timeout_set()
Some internal test usually take 800-900s which is almost
half of the whole sanityn test suite run time. 99.(9)% of
the time the tests just wait to ensure specific order the
operations execute in.

the patch changes cfs_fail_timeout_set() so that it can
interrupt waiting if fail_loc is set to 0 - polling with
1/10s frequency is used.

the tests itself are modified to reset fail_loc. to be
able to do so both operations (referenced as OP1 and OP2
in the tests) are run in background. once started and then
ensured with pdo_sched() helper that MDS threads got to the
blocking points, we can interrupt OP1 and do usual checks.

ONLY=40-47 sh sanityn.sh take: 1017s before and 78s after.

WC-bug-id: https://jira.whamcloud.com/browse/LU-2233
Lustre-commit: 743b85a32e24 ("LU-2233 tests: improve tests sanityn/40-47")
Signed-off-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/4392
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Mike Pershin <mpershin@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>