lustre: llite: move agl_thread cleanup out of thread.
When we start a thread with kthread_create() and later stop
it with kthread_stop(), there is no guarantee that the thread
function runs at all.  So it is not safe to leave cleanup
to the thread.

So move the cleanup code to a separate function which
stops the thread and then cleans up.

Fixes: c044fb0f835c ("staging: lustre: remove 'ptlrpc_thread usage' for sai_agl_thread")
Signed-off-by: NeilBrown <neilb@suse.de>
Reviewed-by: James Simmons <jsimmons@infradaed.org>