lustre: ldlm: fix to stop iterating tree early in ldlm_kms_shift_cb()
It is very possible that old_kms is equal (last extent end + 1),
and if we fail last one check, we will have to iterate all PR extent locks
in tree which is O(N).

This is very likely to happen in IO500 hard mode, because firstly
we try to write which generates a lot of PW locks, after that read
is triggered which need cancel all PW locks to PR locks.

Be careful to avoid overflow, as @l_extent.end could be OBD_OBJECT_EOF,
if that is case, it just means we could stop iterating.

WC-bug-id: https://jira.whamcloud.com/browse/LU-13312
Lustre-commit: b28b3bd9094e ("LU-13312 ldlm: fix to stop iterating tree early in ldlm_kms_shift_cb()")
Signed-off-by: Li Dongyang <dongyangli@ddn.com>
Signed-off-by: Wang Shilong <wshilong@ddn.com>
Reviewed-on: https://review.whamcloud.com/37762
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Gu Zheng <gzheng@ddn.com>

Signed-off-by: NeilBrown <neilb@suse.de>