lustre: fld: convert cache_flush file to LPROC_SEQ_FOPS
Commit 827650494fbe ("staging/lustre/fld: move all files from procfs to
debugfs") converted files to debugfs, but missed the opportunity to
provide a simpler implementation for cache_flush.  When that patch was
taken to OpenSFS lustre in Commit 6d2b66d22c25 ("LU-8066 fld: move all
files from procfs to debugfs"), that opportunity was taken.

So we now copy that improvement back to Linux.

Signed-off-by: NeilBrown <neilb@suse.de>