lustre: llite: remove LOCKAHEAD_OLD compatibility
The OBD_CONNECT_LOCKAHEAD_OLD feature was added for a short time for
compatibility with an implementation of LDLM lockahead that was later
replaced by OBD_CONNECT_LOCKAHEAD2.

Remove the compatibility code for this old implementation that has
been disabled since the 2.13 release.

Also remove other obsolete compatibility code dating back to 2.8.53.

WC-bug-id: https://jira.whamcloud.com/browse/LU-6179
Lustre-commit: 5315db3f1066 ("LU-6179 llite: remove LOCKAHEAD_OLD compatibility")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/38109
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>

Signed-off-by: NeilBrown <neilb@suse.de>