lustre: lnet: handle no discovery flag
When a peer is being discovered, it could indicate that its
discovery is disabled. In this case we shouldn't flag the
peer for deletion. This operation is reserved for cases when
the peer initiates a push to tell us that it has changed
its status from discovery to no discovery.

Fixes: be3f17594603 ("lustre: lnet: advertise discovery when toggled")
WC-bug-id: https://jira.whamcloud.com/browse/LU-12312
Lustre-commit: a83c820f89b3 ("LU-12312 lnet: handle no discovery flag")
Signed-off-by: Amir Shehata <ashehata@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/38229
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>

Signed-off-by: NeilBrown <neilb@suse.de>