gnubee: support u-boot hacking.
The GnuBee can leave the spi-flash in hi-bank mode where reads using
3-byte address (as performed by MIPS for boot) read from the second
16Meg.

If we place an alternate u-boot image at 16Meg, we can cause it to run
on 'reboot', while still having the tried-and-tested uboot there to run
on power-on.  This makes it easy to safely test a new u-boot.

This patch adds a device-tree option to enable "invert the flash bank
on reboot" and adds a new dtb which includes this option and provide
suitable paritioning for the second 16MB - mirroring the first 16MB.

To use this you need a firmware image less than 16MB.

Signed-off-by: NeilBrown <neil@brown.name>