lustre: obd: express timeouts in timeout_t
In the latest kernels time_t has been removed since time_t has
been a 64 bit value just like time64_t so no need for it anymore.
To avoid confusion between timestamps and timeout values Lustre
has a timeout_t typedef which is in seconds and is a s32 since
timeouts are generally short. This helps to avoid errors that
has happens in the past with certain math operation between
timeouts and timestamps that lead to overflow and underflow
cases.

WC-bug-id: https://jira.whamcloud.com/browse/LU-9019
Lustre-commit: 6c6f247bc08b ("LU-9019 obd: express timeouts in timeout_t")
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/31576
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>

Signed-off-by: NeilBrown <neilb@suse.de>