lustre: (re)introduce cfs_curproc_cap_pack() and cfs_cap_t
The set of capabilities used in the lustre protocol exactly matches the
first 32 for the linux kernel capability, but it would be wrong to
assume this will always be the case.
So introduce cfs_cap_t to show where the lustre-capability are being used,
and cfs_curproc_cap_pack() which converts from current_cap()
to the cfs_cap_t sent in the protocol.

Signed-off-by: NeilBrown <neilb@suse.de>