staging: mt7621-dma: remove driver from tree
This driver was a added a time ago and there is no special effort or work
to clean it and any progress at all to mainline it. Also, there is already
a driver from Mediatek in 'drivers/dma/mediatek/mtk-hsdma.c' which looks
pretty similar to this. So the way of supporting HSDMA for mt7621 would be
to modify that driver instead.

Signed-off-by: Sergio Paracuellos <sergio.paracuellos@gmail.com>