mips: ralink: mt7621: add gnubee device trees into the kernel
Device tree blobs for ralink mt7621 SoCs depends on configuration
CONFIG_SOC_MT7621 kernel option. This blobs needs to be properly
builtin into the kernel in order to be able to properly boot the
Gnubee boards. Hence, we need to define specific DTB config options
to know which device tree must be compiled and embeeded into the
kernel for every single board to allow the generated kernel to boot
properly.

Fixes: 7a6ee0bbab25 ("mips: dts: ralink: add MT7621 SoC")
Signed-off-by: Sergio Paracuellos <sergio.paracuellos@gmail.com>