mt7621.dts: restore ethernet reliability.
With this "pause" setting the ethernet often works and sometimes doesn't
- maybe 1 in 4 boots on average.

The symptom is that ethblack doesn't work (eth1 does) and after a few
minutes the network connected to ethblack goes dead.  I think the
switch I have connected to ethblack is confused by something that
ethblack sends it.

Without the "pause" it all seems to work reliably (10 boots, no errors).

Signed-off-by: NeilBrown <neil@brown.name>