nfsd: introduce __fh_verify which takes explicit net arg
This is a step towards having an interface like fh_verify() which
doesn't require a struct svc_rqst *, but instead takes the specific
parts of that which are actually needed.

This first step allows the 'struct nfsd_net *' to be passed in
separately. The new __fh_verify() does not use SVC_NET(), nor does the
functions it call.

The trace calls at the start and end of fh_verify() remain in
fh_verify().  Other callers of __fh_verify() might do their own tracing
with does not require a non-NULL rqstp.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Mike Snitzer <snitzer@kernel.org>