nfsd: __fh_verify now treats NULL rqstp as a trusted connection
The final places where __fh_verify unconditionally dereferences rqstp
involve checking if the connection is suitably secure.  They look at
rqstp->rq_xprt which is not meaningful in the target use case of
"localio" NFS in which the client talks directly to the local server.

So have these always succeed when rqstp is NULL.

With this it is safe to call __fh_verify with a NULL rqstp.
net, cred, client must not be NULL.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Mike Snitzer <snitzer@kernel.org>