VFS: Add ability to exclusively lock any dentry
d_alloc_parallel() already locks an in-lookup dentry before returning
it, and dentry_unlock() will unlock it.

This patch extends dentry locking to all hashed dentries and adds
dentry_lock() which waits for a lock, and dentry_trylock() which locks
only if it doesn't need to wait.  dentry_lock_two() locks two dentries
for rename.

It is planned for this to be used to provide name-based exclusive access
to create (mkdir, mknod, symlink, link, create) a name, to remove
(unlink, rmdir) a name, to rename between 2 names (rename) or to open
(atomic_open) a name.  This will allow i_rwsem to largely be deprecated
for directories, though individual filesystems may still choose to use
it.

When given an in-lookup dentry, dentry_lock() and dentry_trylock() will
immediately succeed as the lock is already held by *this* task.  This
means that d_alloc_parallel() can safely be followed by dentry_lock() to
get a locked dentry.

Once the exclusive lock is obtained on the dentry we must make sure it
wasn't unlinked or renamed while we slept.  If it was, dentry_lock()
will fail so the (in-cache) lookup can be repeated.

With this patch the locking interfaces are introduced but we don't
actually take the lock anywhere new.

There are 3 lockdep locking classes:
 1/ DLOCK_NORMAL context, no other dentry locks held
 2/ DLOCK_RENAME when locking the second name for rename.
 3/ DLOCK_DYING_WAIT when waiting in rmdir_lock() for a child of a
    locked parent to be unlocked.

Signed-off-by: NeilBrown <neil@brown.name>