VFS: Introduce S_DYING which warns that S_DEAD might follow.
Once we support directory operations (e.g. create) without requiring the
parent to be locked, the current practice locking a directory while
processing rmdir() or similar will not be sufficient to wait for
operations to complete and to block further operations.

This patch introduced a new inode flag S_DYING.  It indicates that
a rmdir or similar is being processed and new directory operations must
not commence in the directory.  They should not abort either as the
rmdir might fail - instead they should block.  They can do this by
waiting for S_DYING to be cleared (wait_var_event()).

A new interface rmdir_lock() locks the inode and sets this flag.  It
waits for any d_in_lookup() children to complete and will be extended to
also wait for any locked dentries in the directory, once dentry locking
is supported.

It should be called before attempting to delete the directory or set
S_DEAD.  Matching rmdir_unlock() clears the flag and unlocks the inode.
d_wait_lookup() is no longer static (though it isn't EXPORTed) so that
it is available for use in rmdir_lock().

d_alloc_parallel() is changed to block while this flag is set and to
fail if the parent IS_DEADDIR().  So callers of d_alloc_parallel() no
longer need to test IS_DEADDIR().

rmdir_lock() walks the child/sibling list from the target directory and
so can find dentries that are currently in d_alloc_parallel() and are
in-lookup but not yet hashed in in_lookup_hashtable.  It does not need
to wait for these as d_alloc_parallel() will not complete while S_DYING
is set.  To allow rmdir_lock() to detect dentries in that state we
initialise ->d_in_lookup_hash early.  A memory barrier between the
setting of ->d_in_lookup_hash and testing S_DYING in d_alloc_parallel()
ensures that if it runs before S_DYING is set, then rmdir_lock() will
certainly see that this dentry is hashed, and will wait for it.

Note that while we lock the directory inode to set S_DYING we don't keep
it locked while waiting for d_in_lookup() dentries to be unlocked, as
the owner of the dentry might itself need to lock the directory.  We
lock it again once all children are unlocked.

This new interlock between in-lookup dentries and rmdir means that nfs
and afs no longer need ->rmdir_sem.

__d_wake_in_lookup_waiters() is moved earlier so that it can called by
__d_alloc_parralel() to wake up rmdir_lock() if it is already waiting
when S_DYING is seen to be set.

d_wait_lookup() now needs a subclass as there are two locking contexts
for waiting on an in-lookup dentry
1/ while d_alloc_parallel() waits for an existing dentry to be unlocked.
   After this, an i_rwsem can be taken
2/ while rmdir_lock() waits for a child dentry to no longer be
   in-lookup.  This is done while the parent directory is locked
   with i_rwsem.  In both cases the i_rwsem is I_MUTEX_PARENT.

Signed-off-by: NeilBrown <neil@brown.name>