VFS: provide alternative to s_vfs_rename_mutex
s_vfs_rename_mutex plays a few different roles in managing renames.

1/ It stablises all ->d_parent pointers in the filesystem so that
   ancestral relationships between the source and target directories
   can be determined and preserved.
2/ It prevents any other cross-directory rename so that if the
   source and target directories do not have an ancestral relationship
   then they can be locked in any order without risk if AB-BA deadlock
   with another rename - as all two-inode locking that don't involve
   cross-directory renames are parent-then-child. i.e. ancestral.
3/ It prevents any other cross-directory rename so that the validation
   of ancestral relationship between the source and target which ensures
   no loop is created remains in effect until the rename has completed
   in the filesystem.

A future patch will introduce a new locking mechanism that will allow
filesystems to support concurrent directory operations where they do not
directly conflict.  The use of s_vfs_rename_mutex reduces the
opportunity for concurrency so this patch introduces a new mechanism.

The new mechanism depends on rename_lock and locks against rename only
those ancestors of the two targets that are different.

A/ rename_lock is used to stablise all ->d_parent pointers.  While it is
   held the nearest common ancestor to both targets is found.  If this
   is either the source or destination dentry an error is reported.
B/ While still holding rename_lock, DCACHE_RENAME_LOCK is set on all
   dentries from the two targets up to, but not including, the common
   ancestor.  If any already have DCACHE_RENAME_LOCK set, we abort
   before setting the flag anywhere, wait for that dentry to lose
   DCACHE_RENAME_LOCK, then try again.

   DCACHE_RENAME_LOCK ensures that the parent of the dentry is not
   changed.  DCACHE_RENAME_LOCK of the whole path up to the common
   ancestor ensures that these paths don't change, so the ancestral
   relationship between the two targets doesn't change.  This allows the
   parents to be locked (in any order) and ensures that no other rename
   can cause this rename to create a path loop.

Waiting for DCACHE_RENAME_LOCK to be cleared uses a single
per-filesystem wait_queue_head.  This is signalled once all
DCACHE_RENAME_LOCK flags have been cleared.  This makes it unlikely that
waiters are woken prematurely and may find another DCACHE_RENAME_LOCK
set and so have to go back to sleep.

An important consideration when clearing the DCACHE_RENAME_LOCK flag is
that we do this from child to parent.  As we walk up we need to be sure
we have a valid reference to each dentry.  We will do so because we have
a counted reference to old_parent and both old_dentry and new_dentry,
one of which must be a child of new_parent.  As none of these
directories are empty they cannot be removed and they cannot be moved
away as long as we continue to hold rename_lock.  So as long as we hold
rename_lock while clearing all the DCACHE_RENAME_LOCK flags we will have
a usable reference through the chain of ->d_parent links.  As soon as
rename_lock is dropped the ancestor and everything below that we do not
have an explicit counted reference on will become inaccessible.

Signed-off-by: NeilBrown <neil@brown.name>